Move `Vertices` to its own module
