+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/vertices.rs
@@ -1,0 +1,51 @@
+use serde::Serialize;
+use utoipa::{openapi, ToSchema};
+
+use crate::{
+    knowledge::Entity,
+    ontology::{DataTypeWithMetadata, EntityTypeWithMetadata, PropertyTypeWithMetadata},
+};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "camelCase")]
+#[serde(tag = "kind", content = "inner")]
+pub enum Vertex {
+    DataType(DataTypeWithMetadata),
+    PropertyType(PropertyTypeWithMetadata),
+    EntityType(EntityTypeWithMetadata),
+    Entity(Entity),
+}
+
+// WARNING: This MUST be kept up to date with the enum names and serde attribute, as utoipa does
+// not currently support adjacently tagged enums so we must roll our own:
+// https://github.com/juhaku/utoipa/issues/219
+impl ToSchema for Vertex {
+    fn schema() -> openapi::Schema {
+        let mut builder =
+            openapi::OneOfBuilder::new().discriminator(Some(openapi::Discriminator::new("kind")));
+
+        for (kind, schema) in [
+            ("dataType", DataTypeWithMetadata::schema()),
+            ("propertyType", PropertyTypeWithMetadata::schema()),
+            ("entityType", EntityTypeWithMetadata::schema()),
+            ("entity", Entity::schema()),
+        ] {
+            builder = builder.item(
+                openapi::ObjectBuilder::new()
+                    .property(
+                        "kind",
+                        // Apparently OpenAPI doesn't support const values, the best you can do is
+                        // an enum with one option
+                        openapi::Schema::from(
+                            openapi::ObjectBuilder::new().enum_values(Some([kind])),
+                        ),
+                    )
+                    .required("kind")
+                    .property("inner", schema)
+                    .required("inner"),
+            );
+        }
+
+        builder.into()
+    }
+}
